+++ package.json
@@ -16,8 +16,8 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.18",
-    "@mui/material": "^5.14.18",
+    "@mui/icons-material": "^5.15.1",
+    "@mui/material": "^5.15.1",
     "@types/node": "^20.10.4",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
