+++ src/main.tsx
@@ -1,19 +1,12 @@
 import * as ReactDOM from "react-dom/client";
-import { ThemeProvider } from "@emotion/react";
-import { CssBaseline } from "@mui/material";
-import { theme } from "./theme/theme";
-import { BrowserRouter } from "react-router-dom";
-import ColorModeContextProvider from "./theme/Provider";
-
+import MuiProvider from "./provider/mui/MuiProvider";
+import ColorModeContextProvider from "@/provider/darkmode/DarkmodeProvider";
 import Router from "./router";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <ColorModeContextProvider>
-    <ThemeProvider theme={theme}>
-      <BrowserRouter>
+    <MuiProvider>
         <Router />
-      </BrowserRouter>
-      <CssBaseline />
-    </ThemeProvider>
+    </MuiProvider>
   </ColorModeContextProvider>
 );
